lab6/requirements.txt
pytesseract
pdf2image
Pillow
mysql-connector-python